backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2
django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
pyjwt==2.6.0
pymongo==4.3.2
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
